Add files via upload
Save Sudan is an interactive web page that shows the map of Sudan and some major cities that use renewable data to determine the safety or danger of the city. "The importance of the project lies in its assistance to people to determine the safe areas to help them return / displace to a safer place and the arrival of aid.